trying to add tests to the ai project

+++ MockChessEngine/BitBoard.h
@@ -29,7 +29,7 @@ const BitBoard RANK_8 = RANK_1 << (8 * 7);
 const BitBoard BLACK_SQUARES = 0xaa55aa55aa55aa55ULL;
 const BitBoard WHITE_SQUARES = 0x55aa55aa55aa55aaULL;
 
-enum Square : uint8_t {
+enum Square : int {
 	A1, B1, C1, D1, E1, F1, G1, H1,
 	A2, B2, C2, D2, E2, F2, G2, H2,
 	A3, B3, C3, D3, E3, F3, G3, H3,
